Move util functions to new util module
