+++ wfdb/plot/plot.py
@@ -1,10 +1,10 @@
-import numpy as np
 import os
-import pdb
+
+import numpy as np
 
 from wfdb.io.record import Record, rdrecord
 from wfdb.io._header import float_types
-from wfdb.io._signal import downround, upround
+from wfdb.io.util import downround, upround
 from wfdb.io.annotation import Annotation
